requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
lxml
openpyxl
html2image
python-dateutil
holidays
pytest
